{{cookiecutter.django_name}}/requirements/tests.txt
-r base.txt

coverage
flake8
model_bakery
pytest-cov
pytest-django
tox